Use mmap to load spline coefficients from disk.
Use -w to write a coefficients file.  Use -R to read the coefficients from that
file into memory.  Use -m to memory-map the coefficients from that file.

The idea is to save on memory usage for the spline coefficients (at the cost of compute time).

No checking has been enabled yet.