Use global arrays for coefficients.
Add a flag (alloc_coefs) to the allocation/setup routines to control whether
the normal coefficient storage is allocated or not.
Add a flag (set_local_coefs) to the einspline_spo setup to control the
previous flag and whether or not the values are set.
  For the reference version (using in check_spo and spo_kernel), this should be 'true'.
  For the global arrays version, this should be 'false'.

There is also a define in src/Numeics/Spline2/MultiSpline.hpp, USE_Z_PADDING.
If defined, the z-direction may contain extra padding to ensure alignment (the reference
version stores the coefficients in memory this way)
If not defined, there is no extra padding in the z-direction (and the stride in that
direction is set to the number of splines).

Make sure the number of OpenMP threads is 1. Doesn't work for > 1 thread, yet.