Add unit test framework and one test.
This uses the Catch2 framework, from https://github.com/catchorg/Catch2
and downloading the latest single header.  It is placed in src/Utilities/unit.

It is likely that all the unit tests wil be placed in a single executable
(bin/unit_test).  This is unlike QMCPACK, which has a separate unit test executable
for each top level source directory.